fix(Yggdrasil): use integers for agent configuration

+++ ZalithLauncher/src/main/java/com/movtery/zalithlauncher/game/account/otherserver/models/AuthRequest.kt
@@ -21,6 +21,6 @@ class AuthRequest(
         @SerialName("name")
         var name: String,
         @SerialName("version")
-        var version: Double
+        var version: Int
     )
 }
